apps/auth_app/serializers.py
# Serializer for login credentials — validates email/username and password fields.

from rest_framework import serializers


class LoginSerializer(serializers.Serializer):
    username = serializers.CharField()
    password = serializers.CharField(write_only=True)
