requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
cloudinary
openai
python-dotenv
gunicorn
dj-database-url
